introduce deep cleaning and clean dotfiles in dot

+++ clean.sh
@@ -1,18 +1,16 @@
 #!/usr/bin/env bash
 set -eo pipefail -ux
 
-# bash
+# TODO: these were last removed in september 2026; check if any reappeared
 
-rm -f ~/.{bash_history,bash_logout,bash_profile,bashrc}
+# rm -rf "$XDG_CACHE_HOME"/{appstream,cmp}/
+# rm -rf "$XDG_CONFIG_HOME"/{fsh,github-copilot}/
+# rm -rf "$XDG_DATA_HOME"/man/
+# rm -rf "$XDG_STATE_HOME"/{.copilot,pipewire}/
 
 # bluetooth
 
 rm -rf "$XDG_CACHE_HOME"/obexd/
-
-# claude
-
-rm -rf "$XDG_CACHE_HOME"/claude-cli-nodejs/
-rm -rf "$XDG_CONFIG_HOME"/anthropic/
 
 # electron
 
@@ -24,76 +22,61 @@ rm -rf "$XDG_CONFIG_HOME"/Electron/
 rm -rf "$XDG_CACHE_HOME"/fontconfig/
 rm -rf "$XDG_CONFIG_HOME"/fontconfig/
 
-# github
-
-rm -rf "$XDG_CACHE_HOME"/gh/
-
 # gnome
-
-# TODO recreated on login: all evolution, glycin, gstreamer-1.0, libgetweather, tracker3, goa-1.0, gsd-keyboard..., input-sources..., session.vdb, update-check-LATEST, gvfs-metadata, sounds (so leave autostart as well?), session in state dir
-# application_state to gnome.sh?
 
 rm -rf ~/.gnome/
 
-rm -rf "$XDG_CACHE_HOME"/{evolution,glycin,gnome-calculator,gnome-desktop-thumbnailer,libgweather,thumbnails,tracker3}/
+rm -rf "$XDG_CACHE_HOME"/{gnome-calculator,gnome-desktop-thumbnailer,thumbnails}/
 rm -f "$XDG_CACHE_HOME"/event-sound-cache*
 
-rm -rf "$XDG_CONFIG_HOME"/{evolution,gnome-control-center,gnome-session,goa-1.0,nautilus}
+rm -rf "$XDG_CONFIG_HOME"/{gnome-control-center,gnome-session,nautilus}
 
-rm -f "$XDG_CONFIG_HOME"/.gsd-keyboard.settings-ported
-
-rm -rf "$XDG_DATA_HOME"/{evolution,gnome-settings-daemon,gvfs-metadata,nautilus,sounds}/
-rm -f "$XDG_DATA_HOME"/gnome-shell/{application_state,session.gvdb,update-check-*}
+rm -f "$XDG_DATA_HOME"/gnome-shell/{application_state,gnome-overrides-migrated,update-check-*}
+rm -rf "$XDG_DATA_HOME"/nautilus/
 rm -f "$XDG_DATA_HOME"/recently-used.xbel
 
-rm -f "$XDG_STATE_HOME"/gnome-session@gnome.state
+if [[ ${1:-} == 'deep' ]]; then
 
-# go
+  rm -rf "$XDG_CACHE_HOME"/{evolution,glycin,gstreamer-1.0,libgetweather,tracker3}
 
-rm -rf "$XDG_CACHE_HOME"/{goimports,gopls}/
+  rm -rf "$XDG_CONFIG_HOME"/{evolution,goa-1.0}
+  rm -f "$XDG_CONFIG_HOME"/.gsd-keyboard.settings-ported
 
-# gopass
+  rm -rf "$XDG_DATA_HOME"/{evolution,gnome-settings-daemon,gvfs-metadata,sounds}
+  rm -f "$XDG_DATA_HOME"/gnome-shell/session.gvdb
 
-rm -rf "$XDG_CACHE_HOME"/gopass/
+  rm -f "$XDG_STATE_HOME"/gnome-session@gnome.state
 
-# gstreamer
-
-# rm -rf "$XDG_CACHE_HOME"/gstreamer-1.0/
+fi
 
 # gtk
 
 rm -rf "$XDG_CACHE_HOME"/gtk-4.0/
+rm -rf "$XDG_CONFIG_HOME"/gtk-4.0/
 
-# ime
+# ibus
 
-# rm -rf "$XDG_CACHE_HOME"/ibus/
+if [[ ${1:-} == 'deep' ]]; then
+  rm -rf "$XDG_CACHE_HOME"/ibus/
+  rm -rf "$XDG_CONFIG_HOME"/ibus/
+fi
 
 # lact
 
-rm -rf "$XDG_CONFIG_HOME"/lact/
+if [[ ${1:-} == 'deep' ]]; then
+  rm -rf "$XDG_CONFIG_HOME"/lact/
+fi
 
-# linecast
+# mime
 
-rm -rf "$XDG_CACHE_HOME"/linecast/
-
-# node
-
-rm -f ~/.yarnrc
-rm -rf ~/.{npm,yarn}/
-rm -rf "$XDG_CACHE_HOME"/{js-v8flags,node,node-gyp,yarn}/
+rm -f "$XDG_DATA_HOME"/mimeapps.list
 
 # nvidia
 
-rm -f ~/.nvidia-settings-rc
 rm -rf ~/.nv/
-
+rm -f ~/.nvidia-settings-rc
 rm -rf "$XDG_CACHE_HOME"/{mesa_shader_cache,mesa_shader_cache_db,nv,nvidia}/
-
 rm -rf "$XDG_DATA_HOME"/nvidia-settings/
-
-# nvim
-
-rm -rf "$XDG_CACHE_HOME"/{luarocks,nvim,tree-sitter}/
 
 # qt
 
@@ -101,37 +84,18 @@ rm -rf "$XDG_CONFIG_HOME"/QtProject.conf/
 
 # pulseaudio
 
-# rm -rf "$XDG_CONFIG_HOME"/pulse/
-
-# tensaku
-
-rm -rf "$XDG_CACHE_HOME"/tensaku/
-
-# vscode
-
-rm -rf "$XDG_CONFIG_HOME"/copilot/
-
-# wget
-
-rm -f ~/.wget-hsts
+if [[ ${1:-} == 'deep' ]]; then
+  rm -rf "$XDG_CONFIG_HOME"/pulse/
+fi
 
 # wireplumber
 
-rm -rf "$XDG_STATE_HOME"/wireplumber/
+rm -rf "$XDG_STATE_HOME"/wireplumber/ && mkdir -p "$XDG_STATE_HOME"/wireplumber/
 
 # xdg
 
 rm -rf "$XDG_CONFIG_HOME"/autostart/
 
-# zed
+# dot
 
-rm -rf "$XDG_CACHE_HOME"/zed/
-
-# zsh
-
-rm -f ~/.zshrc
-# rm -rf "$XDG_CACHE_HOME"/gitstatus/ # powerlevel10k
-
-# root
-
-sudo find /root -mindepth 1 -delete
+~/code/dot/clean.sh
